fix: aggregates to the stories too
